WebSockets/tests/submissions/Opera/cookies/003.html

<!doctype html>
<title>WebSockets: sending HttpOnly cookies in ws request</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>

var timeout = setTimeout(function() {
  deleteCookie(false, 'timed out');
}, 10000);
var iframe = document.createElement('iframe');
iframe.src = 'support/set-cookie.php?'+encodeURIComponent('ws_test=test; Path=/; HttpOnly');
iframe.onload = runTest;
function runTest() {
  try {
    var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/ws/echo-cookie');
    ws.onmessage = function(e) {
      ws.onclose = debug;
      ws.close();
      deleteCookie(/ws_test=test/.test(e.data), e.data);
    }
    ws.onopen = debug;
    ws.onclose = function() { deleteCookie(false, e.type); }
  } catch(e) {
    deleteCookie(false, e);
  }
}
function deleteCookie(testPassed, msg) {
  if (arguments.callee.done)
    return;
  arguments.callee.done = true;
  clearTimeout(timeout);
  setTestTimeout(10000);
  // remove cookie
  iframe.src = 'support/set-cookie.php?'+encodeURIComponent('ws_test=; Path=/; HttpOnly; Expires=Sun, 06 Nov 1994 08:49:37 GMT');
  iframe.onload = function() {
    assert_equals(testPassed, true, msg);
    end();
  }
}
document.body.appendChild(iframe);
</script>
